backend/pyproject.toml
[tool.poetry]
name = "mas-backend"
version = "0.1.0"
description = "Multi-Agent System backend with FastAPI and Redis Streams"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = { extras = ["standard"], version = "^0.32.1" }
redis = "^5.2.1"
langgraph = "^0.2.45"
langchain = "^0.3.14"
langchain-openai = "^0.2.14"
google-generativeai = "^0.8.3"
httpx = "^0.27.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
orjson = "^3.10.12"
structlog = "^24.4.0"
celery = "^5.4.0"
SQLAlchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.4"
black = "^24.10.0"
mypy = "^1.13.0"
types-redis = "^4.6.0.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
